+++ src/economy/EconomyState.h
@@ -4,14 +4,14 @@
 
 class EconomyState final: public EventEntity {
   private:
-    long goldAmount = 0;
+    int goldAmount = 0;
 
   public:
     EconomyState() = delete;
-    explicit EconomyState(long initialGold);
+    explicit EconomyState(int initialGold);
     ~EconomyState() override = default;
 
-    void               addGold(long amount);
-    [[nodiscard]] bool spendGold(long amount); // iif goldAmount >= amount: true else false
-    [[nodiscard]] long getGold() const;
+    void               addGold(int amount);
+    [[nodiscard]] bool spendGold(int amount); // iif goldAmount >= amount: true else false
+    [[nodiscard]] int  getGold() const;
 };
